ntp: return error if NTP is not installed
Normally we have at least ntpdate installed, but it is theoretically
possible for neither ntpdate nor ntpd to be installed.

In that case, we should disable the option for network time syncing in
the UI.

We report the lack of an available ntp mechanism as a failure to find
the ntpd.service unit.  A future version of timedated will pick this up
and use it to set the CanNTP property on the interface it exports.

In order to do this, we rework the way gsd-datetime-mechanism-debian.c
is structured so that we can do a (quick) lookup for the availability
of an NTP mechanism without doing a (slower) query if it's actually
enabled.