install ntp-units.d file
Older timedated used to assume 'ntpd.service' if it could not find any
files in ntp-units.d, so we implemented this unit file in systemd-shim.

Newer timedated treats the absence of files in this directory as an
indication that no ntp service is available.  Let's install a file there
to get things working again.