Add a 'power' unit
This unit responds to logind's attempts to poweroff, reboot, hibernate
and suspend.

Suspend and hibernate are implemented via pm-utils if it is installed or
via the kernel directly otherwise.

Take care to note when a reboot or shutdown is requested and ignore
suspend/hibernate requests that follow this (to avoid the problem of
suspending when someone closes the lid after initiating a poweroff).

We also introduce a bit of a dirty hack here: if a suspend is requested
within 1 second of waking up from the previous suspend, ignore it.  This
should help deal with the "double suspend" issues we sometimes see.