Flush the bus before exit
This fixes a theoretical race condition that became less theoretical
with time.

When coupled with the inactivity timeout, it means that the process had
10 seconds to send the replies to method calls from the D-Bus worker
thread before the process would exit and the reply would be lost.  10
seconds is enough for anybody, right?

Unfortunately, the timeout was scheduled from the time each job was
started -- not ended.  Some jobs take longer than 10 seconds and
therefore systemd-shim was quitting immediately upon return to the
mainloop -- immediately after sending the reply to the worker.