Sync the dbus policy with the systemd upstream policy
The dbus security policy for systemd-shim currently differs from that of
systemd upstream.  This difference is unnecessary; the security policy
should be the same whether or not all the listed interfaces are
implemented, avoiding any problems when both policy files are on the
system and systemd is running as pid1.

Bug-Debian: http://bugs.debian.org/746242