cgmanager: stop doing async calls
We need to ensure that the cgmanager actually completes the cgroups
operations before we return to our own caller.  If we don't do that, we
can have races where our caller assumes that everything has been
properly setup, but it actually isn't.

We could probably do this in a more fancy way, but it would involve
private main contexts and counters.  Let's follow the principle of
dumb-is-better in this case.  This code isn't run once per second or
anything...