Implement Abandon and StopUnit for cgroups
Use GDBus subtree support to export a tree of unit files.  Allow Abandon
to be called on these.  Use cgmanager to get a list of cgroups on the
system in order to populate this tree, in case someone tries to
introspect it.

Implement Abandon for cgroup units by calling Prune on the cgroup.  We no
longer call RemoveOnEmpty when the group is first created.

Implement StopUnit for cgroup units by killing all of the processes in
the group and trying to remove it.