lookup_unit: take a string instead of a GVariant
It was convenient for lookup_unit to take a GVariant instead of a unit
name since the only time we wanted to fetch a unit was in the case that
we were handling a message that had the name of the parameter as the
first parameter.

We are about to introduce a new use of this API that won't work this
way, however, so remove this hack and do it more sensibly.