security policy: remove clause for Abandon call
I committed this after copying it out of a work-in-progress branch that
was being worked on by Serge.  I didn't realise that it was most likely
only ever intended to be for debugging purposes.

According to Martin and Stéphane this is not needed and is a potential
security problem.  Let's close that hole.