Send UnitRemoved signal
If a unit gets removed, send out an UnitRemoved(name) signal. This is important
for logind to clean up sessions after closing them.

Note that this is not sufficient by itself: Nothing currently sends the
StopUnit call (this needs to be triggered by a cgroup's notify_on_release
handler).

Prerequisite for https://bugs.debian.org/756076