Implement org.freedesktop.systemd1.Unit ActiveState property
This is e. g. called by logind when receiving an UnitRemoved signal to get the
session scope status. Hardcode a value of "inactive" which is probably wrong
for other use cases, but right for logind session clean up. Proper state
keeping would require implementing the status in subclasses, but let's get to
that when we actually need it.